Crate nalgebra_lapack

linear algebra operations for nalgebra matrices using LAPACK.

Functions to compute the singular value decomposition (SVD) and eigensystem are implemented.

Examples

extern crate nalgebra_lapack;
extern crate nalgebra as na;

use nalgebra_lapack::{HasSVD, HasEigensystem};

fn main() {
    // Create an input matrix
    let m = na::DMatrix::from_row_vector(3,5,&[
        -1.01,   0.86,  -4.60,   3.31,  -4.81,
         3.98,   0.53,  -7.04,   5.29,   3.55,
         3.30,   8.26,  -3.89,   8.20,  -1.51]);

    // Now perform SVD
    let (u,s,vt) = m.svd().unwrap();
    println!("u {:?}",u);
    println!("s {:?}",s);
    println!("vt {:?}",vt);

    // Create an input matrix
    let m = na::DMatrix::from_row_vector(2,2,&[
        2.0, 1.0,
        1.0, 2.0]);

    // Now get the eigensystem
    let (vals, vecs) = m.eigensystem().unwrap();
    println!("eigenvalues {:?}",vals);
    println!("eigenvectors {:?}",vecs);

}

Structs

NalgebraLapackError

Traits

HasEigensystem
A type for which eigenvalues and eigenvectors can be computed.

HasSVD
A type for which a singular value decomposition can be computed.

Type Definitions

NalgebraLapackResult