Trait nalgebra_lapack::HasSVD

pub trait HasSVD {
    type VectorType;
    type MatrixType;
    fn svd(self) -> NalgebraLapackResult<(DMatrix<Self::MatrixType>, DVector<Self::VectorType>, DMatrix<Self::MatrixType>)>;
}

A type for which a singular value decomposition can be computed.

Associated Types

type VectorType

type of the output vector element (e.g. f32)

type MatrixType

type of the output matrix element (e.g. f32 or Complex<f32>)

Required Methods

fn svd(self) -> NalgebraLapackResult<(DMatrix<Self::MatrixType>, DVector<Self::VectorType>, DMatrix<Self::MatrixType>)>

svd computes the singular value decomposition (SVD). Returns full matrices.

Because the input matrix may be overwritten or destroyed, it is consumed.

Returns

u - The left-singular vectors.

s - The singular values.

vt - The right-singular vectors.